some crud additions

+++ Backend/RRSS/src/main/java/com/demo/rrss/rrssbackend/service/CategoryService.java
@@ -9,35 +9,55 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.demo.rrss.rrssbackend.entity.Category;
+import com.demo.rrss.rrssbackend.entity.Users;
 import com.demo.rrss.rrssbackend.repository.CategoryRepository;
+import com.demo.rrss.rrssbackend.repository.UsersRepository;
 import com.demo.rrss.rrssbackend.rest.request.AddCategoryRequest;
+import org.springframework.ui.Model;
+
 
 @Service
 public class CategoryService {
 	@Autowired
 	CategoryRepository repository;
-
+	@Autowired
+	UsersRepository usersRepository;
 	public Category getCategory(Long categoryId) {
 		Optional<Category> response = repository.findById(categoryId);
 		System.out.println(response);
 		return response.get();
 	}
 
-	public void addCategory(AddCategoryRequest category) {
+	public void addCategory(AddCategoryRequest category, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		Users existingUser = usersRepository.findById(userId).get();
+		if (!existingUser.getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not authorized to delete a category");
+		}
 		Category newCategory = new Category();
 		newCategory.setCategoryName(category.getCategoryName());
 		newCategory.setDescription(category.getDescription());
 		repository.save(newCategory);
 	}
 
-	public Category updateCategory(Long categoryId, Category categoryDetails) { // TODO yetki kontrolü eklenecek
+	public Category updateCategory(Long categoryId, Category categoryDetails, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		Users existingUser = usersRepository.findById(userId).get();
+		if (!existingUser.getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not authorized to delete a category");
+		}
 		Category category = getCategory(categoryId);
 		category.setCategoryName(categoryDetails.getCategoryName());
 		category.setDescription(categoryDetails.getDescription());
 		return repository.save(category);
 	}
 
-	public void deleteCategory(Long categoryId) { // TODO yetki kontrolü eklenecek
+	public void deleteCategory(Long categoryId, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		Users existingUser = usersRepository.findById(userId).get();
+		if (!existingUser.getIsAdmin()) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not authorized to delete a category");
+		}
 		if (repository.existsById(categoryId)) {
 			repository.deleteById(categoryId);
 		} else {
